Person3/FastAPI/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
duckduckgo-search>=6.0.0
exa-py>=1.0.0
tavily-python>=0.3.0
boto3>=1.34.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
